test: stabilize phase 1-3 quality gates
